+++ libcxx/test/std/language.support/support.limits/support.limits.general/iterator.version.compile.pass.cpp
@@ -20,6 +20,7 @@
                                             201811L [C++20]
     __cpp_lib_constexpr_iterator            201811L [C++20]
     __cpp_lib_make_reverse_iterator         201402L [C++14]
+    __cpp_lib_move_iterator_concept         202207L [C++20]
     __cpp_lib_nonmember_container_access    201411L [C++17]
     __cpp_lib_null_iterators                201304L [C++14]
     __cpp_lib_ranges                        202106L [C++20]
@@ -41,6 +42,10 @@
 
 # ifdef __cpp_lib_make_reverse_iterator
 #   error "__cpp_lib_make_reverse_iterator should not be defined before c++14"
+# endif
+
+# ifdef __cpp_lib_move_iterator_concept
+#   error "__cpp_lib_move_iterator_concept should not be defined before c++20"
 # endif
 
 # ifdef __cpp_lib_nonmember_container_access
@@ -74,6 +79,10 @@
 # endif
 # if __cpp_lib_make_reverse_iterator != 201402L
 #   error "__cpp_lib_make_reverse_iterator should have the value 201402L in c++14"
+# endif
+
+# ifdef __cpp_lib_move_iterator_concept
+#   error "__cpp_lib_move_iterator_concept should not be defined before c++20"
 # endif
 
 # ifdef __cpp_lib_nonmember_container_access
@@ -113,6 +122,10 @@
 # endif
 # if __cpp_lib_make_reverse_iterator != 201402L
 #   error "__cpp_lib_make_reverse_iterator should have the value 201402L in c++17"
+# endif
+
+# ifdef __cpp_lib_move_iterator_concept
+#   error "__cpp_lib_move_iterator_concept should not be defined before c++20"
 # endif
 
 # ifndef __cpp_lib_nonmember_container_access
@@ -158,6 +171,13 @@
 # endif
 # if __cpp_lib_make_reverse_iterator != 201402L
 #   error "__cpp_lib_make_reverse_iterator should have the value 201402L in c++20"
+# endif
+
+# ifndef __cpp_lib_move_iterator_concept
+#   error "__cpp_lib_move_iterator_concept should be defined in c++20"
+# endif
+# if __cpp_lib_move_iterator_concept != 202207L
+#   error "__cpp_lib_move_iterator_concept should have the value 202207L in c++20"
 # endif
 
 # ifndef __cpp_lib_nonmember_container_access
@@ -209,6 +229,13 @@
 # endif
 # if __cpp_lib_make_reverse_iterator != 201402L
 #   error "__cpp_lib_make_reverse_iterator should have the value 201402L in c++2b"
+# endif
+
+# ifndef __cpp_lib_move_iterator_concept
+#   error "__cpp_lib_move_iterator_concept should be defined in c++2b"
+# endif
+# if __cpp_lib_move_iterator_concept != 202207L
+#   error "__cpp_lib_move_iterator_concept should have the value 202207L in c++2b"
 # endif
 
 # ifndef __cpp_lib_nonmember_container_access